SYN-612: Address review of the embedded core configuration
Four fixes, all in code this branch added.

The new tests in runtara-environment's runtime module declared their own
env mutex while the config module already had one. Two mutexes in one
test binary serialize against nothing, so the SAFETY comments were false
and `set_var` really could race a `getenv` on another test thread. Move
the guard to a crate-wide test_env module and have it hold the lock
itself, so there is no second lock to forget. It restores each variable
on drop, panics included, instead of unconditionally removing it.

Correct the core_overrides doc: build() reads and validates the
environment whether or not an embedded core is configured.

Read core's configuration before opening the pool in maybe_start_embedded
so a malformed value is caught before migrations run.

Also correct three doc comments that claimed a malformed value "fails
startup". True of the core and environment binaries; runtara-server
reports the error and continues without an embedded runtime, which is
how it already treats every other embedded-start failure.
